Refactor: create ShareContent for split child component
